Add support for resizing disk on vm setup
The new config option `disk_resize` is passed to `qemu-img resize`
command after the initial image is created. This allows to set or adjust
the image size before starting the vm for the first time.

Resizing is limited to one image only. Error is signalled if
multiple image files are provided by the box or by the config.